Add autofight_warning option.
The goal of this option is to be the counterpart of
tile_key_repeat_delay=0 for console mode.

Since we don't know if the player keeps the tab key down in console
mode, I keep track of time elapsed between two tab commands in
milliseconds (using std::chrono).  If that delta is less than the
autofight_warning option, we stop the user with the message "You should
not fight recklessly!".

The option has no effect if the tab command was issued by a macro.

[Committer's note: Squashed several commits.
 Closes #244 pull request. -nfm]