Don't lose orc corpse identity on save + restore.
We stored the just-killed orc in its corpse for future resurrection.
Unfortunately, dead monsters are marshalled as just MONS_NO_MONSTER (two
bytes) without any of their old information.  So after saving and
restoring, the dead orc had become a "DEAD MONSTER".

Now restore the corpse's copy of the orc to full HP so that it persists
across a save.  Furthermore, ensure that it has at least 1 maxhp, just
in case (if not, resurrection would put a DOA monster in the grid, with
a potential crash when something walks into it).

Existing corpses containing a DEAD MONSTER are cleaned out on restore.
Unfortunately this makes them non-resurrectable, but all the necessary
information has already been lost, so there's not much that can be done
about that.