zot:5 side-buff(?)
After cleaning up a very nasty header and leaving some notes, an old and
questionable status is changed. Seventeen fixed spots in the chamber halves
closer to the orb have three separate equally-likely arrangements:

Randomly, per spot |   A    |   B    |   C
---------------------------------------------
Ancient Lich       | 50.00% | 16.67% | 16.67%
Orb of Fire        | 16.67% | 50.00% | 16.67%
Random floor spawn | 16.67% | 16.67% | 16.67%
Chance for a trap  | 16.67% | 16.67% | 50.00%

Option C seems like an unusual wasted opportunity, a shift away from plenty
of monster threats high monster threat onto the not-nearly-as-dangerous extra
traps. As such, in a slight shift of balance, I'm changing the setting to the
below chances, of more random floor spawns and killer klown abuse:

Randomly, per spot |   C
---------------------------
Ancient Lich       | 16.67%
Orb of Fire        | 16.67%
Random floor spawn | 25.00%
Chance for a trap  | 16.67%
Killer klown       | 25.00%

As most random floor spawns and killer klowns don't quite hew towards the
nastiness of many orbs of fire or ancient liches (short of, perhaps, a rather
lucky moth), this is not exactly a straight buff as much as it ups the
standards of danger. There are plenty of other options for this arrangement,
with so many different branch threats, but these were chosen as one of the
least noticeable of options.