Rename AF_CHAOS to AF_CHAOTIC (giann)
The old name collides with the address family macro for the MIT CHAOS
protocol, causing build errors on cygwin.  We tried to #undefine the
macro in libw32.cc, but it was also being pulled into tileweb.h (through
sys/un.h).  Rather than try to chase down everything that might pull in
a networking header, just rename our attack flavour.

Maybe it would be better to make this a scoped enum and then use
something like attflav::chaos, but that's a significantly bigger change
(and does make the names longer) so I am not doing that right now.