Correct some minor tag checks in tag_read_char.
Two fixes:

  1. The numbers for the explore mode minors (in the post-34 branch of
     the #if) were off by 3.

  2. The version check at the end (for the corrected location of the
     explore mode bool) was incorrect.  If some older (< 34) major tag
     had more than 130 minor tags, we would have gotten a short read
     trying to unmarshall a boolean that wasn't there; fortunately,
     34 was the first major tag with this many minors.  Worse, on a
     major tag bump, we would have tried to unmarshall that bool in
     all old saves.  (Old majors matter here because this is the
     char chunk, which is shown in the save browser even for games
     that are too old to transfer).

And one bit of cleanup: Remove the #if TAG_MAJOR_VERSION around the
first explore mode tag check and just use the numbers, like elsewhere
in the function.  No sense leaving in two versions of the code that
will go out of sync.  Add a comment explaining the source of the
numbers.