Fix another rare Imprison crash
In very specific circumstances, it was possible for Imprison's
validation to conclude that a monster could be shoved out of the
way, but the actual execution of the Imprison to not have space
for that monster - because the logic of the latter didn't
explicitly veto the spaces the walls would soon occupy. This
crudely fixes that; there are probably still some insane corner
cases that this doesn't handle, and it might be better to build
a map of monster -> shove locations rather than reproducing it,
but this should be good enough for now.