Remove lantern of shadows
The primary function of lantern of shadows was to enable tedious kiting,
which was both excessively strong and also boring. The secondary function
of reducing LOS while removing the ability to effectively attack or cast
was never particularly interesting either, but was technically optimal as a
fast way to get monsters at edge of LOS out of LOS.

We already have too many misc items that summon monsters, and this one was
problematic, so it's getting the axe.