Don't let speed brand change mindelay skill (#10333)
The game claims a longbow of speed has mindelay 0.4 (which appears
intended to be a white lie, see 0.18-a0-72-g2733cf4) at skill level 20.
But skill 20 corresponds to an average delay of 0.47, and skill 22 does
in fact reduce the average delay to 0.4.  The intention, though,
appeared to be for the min delay skill to actually be unchanged by the
speed brand.

Instead of capping the speed based on (truncated) weapon min delay, cap
the contribution of skill to the min delay skill level.  Another
possibility would have been capping with weapon_min_delay(*weap, false)
before applying speed brand.