Dramatically upgrade hell hogs from sticky flame to fireball
The hell hog is redundant. It is a silly, flavourful monster, certainly, and
it has its use in Porkolator, but it doesn't nearly make the mark for dangerous
or notable enemies in extended. Even with essentially twice the stats of hell
hounds and decent band sizes, their stats are effectively worse than the
lair branch threat of harpies for a monster that effectively only appears in
extended. With item destruction long gone and extended defenses readily taking
it, sticky flame breath hardly matters either.

Thus, with some rF++ to make it work, hell hogs get a significant buff. The
whole pack of three to five hell hogs now gets to blast players with 3d20
fireballs, at the draconic standard rate of 6/20 actions (on a breath timer).
The high degrees of constant, very loud fire damage on a fast band is a
certainly passable threat for some: they are thus set to appear a bit more
commonly in Gehenna (cutting into derived undead weight) there. They are
certainly no shard shrikes, but it still provides another variation of fire
threats for the branch of damnation (and elsewhere, as adjusted).