Weaken wild/subdued magic
A single level of a 3-level mut had an enormous impact on a
a character - increasing spellpower by 50%, or decreasing it by
33%, is really huge! The change in spell failure chance mitigated
this to a certain extent, but not enough.

This reduces the effects to about 3/5ths of what they were before;
each level of wild magic now increases spellpower by 30%, and
subdued magic now reduces it by 20%. The effects on spell failure
were scaled similarly, with subdued magic's effects on spell
failure being reduced even further, to reinforce its status as a
Bad Mut.