Alter draconian caller spells
Did you know Summon Drakes can summon 10% of the time a variety of standard
dragons? This is, of course, an unusual and awkward arrangement: outside of
this, the spell in question is mostly a flood of space-filling monsters as
encountered in Zot. The best it gets is a ally-unfriendly proxy use of
miasma mixed with noxious clouds, though swamp and death drakes.

Instead of letting draconian callers rarely get some dragons through such a
spell, the dragons are cut from the spell and replaced with lindwurms (well,
"drakes" does seem to include komodo dragons already). They also directly get
access to the stronger, equally thematic and rare employment of the Summon
Dragon monster spell. Not at too often a frequency (cutting into their summon
drake frequency, actually), but enough to see the occasional iron dragon -
a bit of a buff.

Draconian classes are kind of bloated relative to their combination nature,
with some awkward roles and conjurations overlap. The old support-focused
design is unusual for a late-game summoner, but I'm not sure what else should
be done with callers. Probably not giving them summon draconians as the
unused code removed here possibly supports.