Change giant eyeball speed / gaze, grant random band, randomly spawn again
Giant eyeballs are interesting because they are extremely focused in their
support: an irresistable paralysis that merely needs line of sight, set on an
otherwise-harmless, fragile, slow monster. However, their intended design as
universal high priority dwindles away against luring tactics, due to their low
speed, and they couldn't rely on company to threaten with when set to spawn
normally in D or Depths or Slime. As such, they entered an awkward purgatory,
seen rarely in unusual vaults or amongst more common eyes with the sprasely
used Summon Eyeballs spell.

These factors are now addressed in several ways. Giant eyeballs are now normal
speed, but the paralysis gaze now uses the sputtercast mechanic (from orb
spiders and salamander stormcallers), requiring three casts to full charge and
paralyze a target (here thus a rough 40% chance to be paralyzed once in 60 aut
in LOS, as opposed to 75% in 66 aut). Alongside the instanteous nature of the
gaze spells, this allows them to follow along with other monsters (and removes
the chance of a giant eyeball's speed versus one's one colliding into instant
paralysis). This is paired with a granted singular RANDOM_MOBILE_MONSTER band
(also provided to eyes of draining), to guarantees a little company to support
with and be supported for that slow but persistant gaze charge. Finally, they
get an HD buff to reduce the chance of monsters blasting through them to get
to the player.

With the sum of these changes hopefully making them less abusable monsters,
these giant eyeballs are re-introduced into the Depths and Slime floor spawn
sets (at pretty low weights). Both branches demonstrate the functionality of
fragile support threats, with boggarts and giant orange brains, and it should
be a fresh new opportunity to assess these retooled eyeballs alongside stone
giants or acid blobs.