Alter Salamander Stormcaller spells, tweak their D placement.
Sticky Flame Range is cut from their spellsets, and the Fire Storm charge
chance is marginally increased. Possessing more than one spell on a monster
defined by a big and fancy and unreliable spell such as that triple seperate
casts to charge up a very early Fire Storm seems distracting for the monster's
actions and in terms of its design - which is already precarious, as it is
stuck on a very unreliable impending (now slightly better) timer of a chance
to cast big conjurations. Sticky Flame is also... not exactly the strongest
of effects by even late D relative to the AF_FIRE and that storm.

The OOD range for Stormcallers is also expanded a bit and weighted down
a bit in return, to permit a wider range of levels and vaults to actually
access and spawn their slim out-of-depth part of the large D out-of-depth
monster list.