Prevent elemental adaptation at max resists
Before, the resistance would apparently trigger and show the rF+ status
indicator, but have no effect -- despite the misleading status
indicator.  With this patch it will silently fail if the player is at
max resists for the relevant element at the time of exposure.  This
could be a very slight buff in that AC can always trigger, and it is now
more likely to be continuously on when receiving hybrid damage (e.g.
brands) when at max resists.  However, AC is the weakest of the resists
at the point in the game where this is typically relevant.

This doesn't do anything with equipment, so you could still get the
status indicator on at max resists by letting it trigger and then
putting on the relevant ring.