Allow armored quarterstaff users (fighter)
If a player wants to start with a quarterstaff and armor (plus a shield they can't wield), let them. They may be planning to transition into an enhancer staff later in their career, or perhaps they just want to lean on the awesome early game power of a quarterstaff while they train up shields and other skills.

A quarterstaff start is a great start, and for some races, an extraneous shield is preferable to being sent down the Gladiator-style light armor route.