Don't make enchant weapon/armour scrolls remove curses
There's no particular reason for them to do that in addition to their primary
effect of increasing enchantment, other than the fact that that's how
enchanting items tends to work in other roguelikes. It just makes them
unnecessarily do an extra thing, when there's already a dedicated scroll for
curse removal.