Slow Iron Giants down a bit
They're the deadliest monster in every area they appear, both in
Depths and all of Hells. This includes Dis specifically, which is
*already* the nastiest hell. It's excessive!

Reduce their speed to 8 to start; they're ranged enemies still, so
this shouldn't be a problem. (It's also nice to have some slow
enemies, for variety.) Possibly their bands should also be adjusted,
especially in Dis.