Remove power cap on Lugonu Banish
The old Lugonu Banish ability hit a power cap at 10.5 Invocations skill. This
changes the ability to use a flat scaling with no cap. The ability is harder
to invoke at low Invo, but has higher base power. Each level of Invo increases
the power by 3, up to a maximum of 149 power at Invo 27.

This change should make Banish slightly worse at 0 Invo, relatively similar
from 1 to 11 Invo, and somewhat better at 11-27 Invo. Banish didn't need
a buff, but it was either this or document that the power caps at 10.5 Invo,
and of the two options, this one seemed better.