nerf ossuary bog body
These things are more dangerous than an ogre, in a portal that players can typically count on being low in danger and poor in loot. Nasty surprise for the unspoiled, especially players with a few wins under their belt who know what to expect from low-level portals.

Still deals 25 damage!!!!!!!!!! but is a lot easier to kill... if it doesn't 1-hit the Spriggan/DE player.