Don't make the Sword of Zonguldrok passively animate hostile dead
Just being a reaping weapon makes it special enough, and animating hostile dead
was unexpected in cases where, for example, enough damage was dealt by aux
attacks rather than the reaping brand for the brand not to trigger on kill.
Also don't make it recurse on equip, since with the brand depending on damage
dealt (and with the weapon no longer enforcing a corpseless conduct) it's fine
to allow swapping.