Drafts to buff, serialize, pair-up, and announce (fixed) demonic rune vaults
The demonic rune is possibly the most awkward of any runes in the game.
Randomly found within any conventional floor of Pan and their vaguely-hostile
Pandemonium Lord vaults or reliably found in a spoiler-beneficial manner from
gimmick vaults of varying lethality or the sprawling sudden shift of the holy
floor, the demonic rune invites canvassing or diving through the limited range
of the monotonously infinite. It is a unique practice, perhaps, broken up by
the more structured intimidating non-random Pan Lords and providing different
approaches between checking every floor or merely whatever one encounters,
but there is little fixed threat or tangible feedback for players in such an
approach. It is one of the most unpleasant parts of the already troubled
realm of extended, but without it there is little point and thus variety to
the levels between the other four Pan runes.

Thus, these new arrangements.

The demonic rune can still be found randomly within the floor vaults of Pan,
for the sake of weak characters or cautious play. Eventually, one will
encounter a serial vault (or the unchanged Holy Pan floor) announced with
"the tension of some great conflict"- that serial vault is pan_highlights,
which places two upgraded guaranteed demonic rune vaults and a few
corresponding flavour minivaults:

  * _draining_boxes, with eyes of draining in grate boxes;
  * _hall_of_pain, with smoke demons and tormentors behind walls of grates,
  * _hellion_isle, with its famous cluster of sixteen hellions behind lava,
  * _frozen_over, with nine blizzard demons behind a wall of ice statues,
  * _disco_hall, promoted up here, with high numbers of strong branch spawns.

Of course, only one of the vaults in the serial vault actually contains
the demonic rune, similiar to the decoy huts of the Shoals' barnacled rune.
The intent of keeping both methods (announced and unannounced) is meant to
provide and preserve choices of danger similar to the multiple floors of Abyss.

---- Implementation Issues ----
 * Rune decoy code, airlifted from the Shoals implementation, seem completely
   non-functional- both main vaults will place the demonic rune.
 * The announcement messages using crawl.mpr repeats for multiple level
   build attempts, which looks very bad- it should probably be built-in.
 * The placement of both main vaults relies on dgn.place_maps with a single
   die_on_error, which rarely results in levels with only one main vault and
   the occasional visible lua errors regarding failed level vault placement.
   What else could be used to make the twin large vaults consistently place?

---- Future plans ----
This is only the start of work and proposals regarding Pandemonium and the
demon branches of extended as a whole. Further reading is available at:
https://crawl.develz.org/wiki/doku.php?id=dcss:brainstorm:dungeon:hell-pan_roulette

Regarding the demonic rune, however:
 * Set-up a second demonic rune serial vault, placing four or five of the normal
   random demonic rune vaults with upgraded random contents (as opposed to
   default floor spawns) and a guarnteed demonic rune all on one floor.
 * Set rune levels (these demonic rune vaults, though possibly also the other
   four fixed lord rune levels) to increase in placement chances with acquired
   runes, slain Pan lords, or both.
 * (Revamp about half of the holy monsters guarding that additional demonic
    rune vault of the holy pan floor, whether or not it gets split off entirely
    as a seperate short rune branch in the Hell-Pan roulette plans.)