Refactor monster shouting
The function name prefix "handle_" means nothing, has never meant
anything, and serves as a sort of filler. It is a Code Smell, and
should be purged whenever it appears.

This commit makes wizmode shout-forcing work even when a monster
would be unable to shout normally; e.g. if they were silenced,
sleeping, etc.