Ukayaw: get this party started
Add Ukayaw, the god of ecstatic dance. Ukayaw has extremely rapid piety
gain and loss in combat as long as you keep the fight going. The god's in-game
description:

Ukayaw the Reveler is a god of ecstatic dance. On the surface, Ukayaw
appreciates many forms of worship, but in the dungeon most worshippers stick to
the dance of combat. Ukayaw appreciates the passion and rhythm of combat,
rewarding followers for each strike they deliver and for the damage they deal
their foes. The longer the dance continues, the greater powers Ukayaw offers.
When the dance ends, Ukayaw quickly loses interest.