Ukayaw: Add *** passive, solo time
Upon reaching ***, Ukayaw briefly paralyzes all enemies on LOS. The duration
is at least 10 AUT. It increases with Invocations skill and the difference
between player level and monster HD, but it takes a significant amount of
those factors to raise the duration significantly. Maximum duration is 40
AUT.

This triggers passively so that some aspects of the 'combo' this god offers
don't require manual intervention. Plus, it'll be a cool event to aim for.

The ability only triggers at most once every so often so that it can't
be abused via repeated piety loss/gain.