Ukayaw: Add **** passive, pain bond
Like the *** power, this triggers automatically upon hitting the required
piety. It gives all monsters in LOS the pain bond status for an amount of
time that increases with Invocations skill. While the status is active,
affected creatures deal a portion of the damage they receive to nearby
creatures with the same status -- which can trigger the status again in a
chain reaction. The damage shared decreases with distance, dealing 80%/60%/
40%/0% at one/two/three/four tiles away, respectively.