Hepliaklqana 3* ability: Transference
Swaps your ancestor with a targetted creature (which can be you!)
if your ancestor is in LOS. At 5*, adjacent enemies are also slowed;
the amount of slowing scales with Invocations. Regardless, costs a
small amount of piety.

Intended to be a fun tactical trick; escape ability, offensive
play, whatever. The sort of thing you can only do with an ally
god, as with Idealise.