Don't allow Pakellas device surge to work on +Blink, +Rage etc
Especially since it actually did nothing but waste MP in those cases.
Thematically it seems fine to not treat evokable randart/ring/amulet properties
as "devices", and just keep it as affecting wands/rods/misc evokers (which are
the things that the player is likely to try and use it on anyway).