Don't omit mention of Dispater's staff, +0 rods (CanOfWorms)
We got the message "Dispater is ." (likewise Cerebov and others).  The
missing predicate occured when several monsters came into view
simultaneously, and one was wielding an unidentified unrandart (e.g. the
staff of Dispater).  In _monster_headsup we decided to give the
announcement because wpn->brand, being shared with the unrand index, is
nonzero; but then we did nothing in get_monster_equipment_desc because
the item was unidentified but we passed DESC_IDENTIFIED.

We also failed to give any announcement for a monster with a +0 rod, or
with a magical staff, because those had wpn->brand == 0.

Instead, pass DESC_WEAPON so the weapon is always described, even if
unidentified; and produce the message if (and only if) the weapon is a
branded weapon (using the correct get_weapon_brand check), an unrand,
a rod, or a staff.  Continue to use DESC_IDENTIFIED for Ash ID, so as
to minimise the potential impact of this change.