Make frenzy dispellable (10450)
It's weird that you can dispel the haste/might from frenzying an
enemy separately from the insanity; it's also weird that you can
dispel confusion, but not frenzy. Of course, with this change, it's
weird that you can dispel frenzy but not berserk... but it'd be
pretty messed up to be able to dispel a bear berserking, or to have
your own berserk dispelled by a purple drac. Weirdness either way,
but this seems like it's probably the best compromise.

It'd be really nice if we communicated dispellability better.