Don't warn when confused-moving near avatars
And other similar monsters. You're not going to anger them, so
it's not a critical enough problem to give a warning prompt for,
and it's quite annoying to have the prompt when it's not needed.

This does remove the prompt when actually moving into an avatar or
ancestor while confused, which will attack them (if successful),
not swap with them... i feel like that's okay, but if not, at
least there's some refactoring here.