Make {autopickup} stash annotations into an option, autopickup_search.
This stash annotation was added in 78e7867c. It has minor utility but has
the potential to slow down the game in various ways. The problem is that
this stash annotation makes the game check autopickup status of every item
in the game whenever you use Ctrl-F and check autopickup status of every
item in view whenever you take any action at all. If you have a complicated
custom autopickup function (which is more popular now than it once was) and
there are a lot of items in the game or in view then this can be quite
expensive and can interfere with other unrelated uses of player lua.

I can see an argument that this option should default to true because for
most players it is safe to use, but I also think the feature can cause
a fair amount of trouble compared to how minor its utility is.