Relax the lua throttle somewhat.
The previous throttle parameters were the same that were chosen when it was
first added nearly 9 years ago - if a player lua call made more than 10000
instructions before returning then there would be a delay of 2 milliseconds,
and then for every additional 10000 instructions that delay would double,
up to a maximum of 1024 ms. After 100 such delays the player lua would be
terminated.

This commit increases 10000 to 50000, allowing player lua to make five times
as many instructions before getting delayed the same amount of time. I don't
believe that this will have a noticeable impact on server performance or
stability - only bots like qw or gw were consistently throttled very much,
and they already imposed a delay of their own in addition to the throttle.
The throttle will also still serve its primary purpose of ending infinite
loops caused via lua.

However, "regular" player lua was beginning to cause unpleasant half-second
or longer delays under certain circumstances because of the throttle. In
particular, recent upgrades to autofight made it be significantly throttled
when a huge number of monsters were in view (e.g. in a late Ziggurat level).
Relaxing the throttle by a factor of five reduces half-second or second
delays to an unnoticeable 6 ms. Of course some scripts will still hit the
throttle and we should still try to avoid unnecessary inefficiency, but
relaxing it a bit seems reasonable.