Use the same deck types for Nemelex gifting and random generation
The distinction between gifted decks (escape, destruction and war) and
non-gifted ones (summoning, defense and changes) was no longer a very
meaningful one and mostly just consisted of some reshuffling of the various
subdecks. The only noticeable difference being that destruction cards weren't a
part of any of the non-gifted decks, which didn't seem like an especially
important restriction.

Removes decks of summoning, changes and defense in favour of letting escape,
destruction and war generate at a normal rate.