Rework Xom's smoke trail.
Turn it from a "bad" effect (which was really more a no-op) to a good
effect that leaves a random cloud trail behind you.

This is considered good because you can use it to cut off monster
pathing and so on, although cutting off your escape is a potential
problem.