Simplify book acq (|amethyst, minmay)
Trim out a check for spell skills being >= 1, which seems to date
to when you needed to read scrolls to be able to cast spells at
all. This makes also makes it possible to get spellcasting manuals
under Trog, which isn't great for troggites, but perhaps compensates
for getting guaranteed manuals.