Remove M_GLOWS (chequers)
Had the following effects:
- Prevented monsters from being turned invisible.
- Prevented monsters from being enchanted with Corona.
- Reduced monster stealth (!?) by 3 points.

Did not have any effect on attack accuracy.

Was poorly communicated to players (many people assumed it *did*
increase accuracy - understandably!), and was almost irrelevant
to play.

Could have been replaced by a new 'perma-corona' status on each
of these monsters, and a compensating buff to EV, but: why? All
of the relevant monsters are interesting enough without some new
UI and mechanical clutter slapped on.