Hexer specialization changes
Paralysis vs Metabolic Englaciation was not a remotely comparable
or interesting choice; the latter was very trivially superior. I
thought this would be the case when I implemented it, but I was low
on ideas and wanted to give it a shot. The shot is now over.

Hexers' Slow is now automatically replaced by Paralyse at XL 15,
and the specialization choice is now between Englaciation and Mass
Confusion at XL 21. The ancestor should still be strong (since both
of those are very strong effects to have!), but somewhat less so,
and hopefully the choice will be less one-sided.

Doesn't seem worth adding save compat for; players who chose
Paralyse before the change will wind up with Mass Confusion after
(once they hit XL 21).