Attempt to reduce player whinging
Count ancestors as avatars for purpose of XP gain (full XP from
ancestor-killed enemies.)

Ancestor/Hep power can be controlled by other, more elegant knobs
than XP gain, and it's not desireable to discourage players from
using their ancestors in fights. The opposite of that, really.
The only reason I didn't make this commit before is because it's
an ugly special case, but hopefully someday we can find a better
way to balance summons and remove the reduced-xp thing for them,
and then the special case will be gone.

Huzzah.