Remove the EV penalty for taking stairs
It's a pretty obscure mechanic that you'll only realise exists if you
watch the EV field of the HUD intently.

To compensate, the delay on a new level after using stairs is increased
from 4/3 turns on an existing level and 2/3 on a new level to 3/2 and
3/4 turns respectively.

This is a spiritual sequel to e1c5091, which removed the penalty for all
delays other than taking stairs.