Personalise the Torment card
To apply holy word to those vulnerable to it, and torment otherwise. Also
rework holy word to be inconsistent but in a different (and possibly better)
way to the way it was previously inconsistent (by just having fixed separate
monster/player effects rather than changing the effect based on the reader,
not that it matters either way since monsters can't read it).

Could maybe also do with a rename, but possibly Torment is still fine (since
you don't see the name of the card when you draw it anyway, and if you look
up the description it's still accurate for either effect).