Control some more vault placements of hells enemies
Boot a few shard shrikes and a doom hound from Depths vaults, to continue to
demarcate them and iron giants as Hells-specific. The shard shrikes in a
slim few Zot vaults are okay, though, because it's Zot.

Also, futz around with the old and badly-arranged _tower_of_babel: it's now an
enclosed stair bubble vault with solely Dis appearances, and it edges closer
to resembling the level it's a homage to, rather than nearly always being
approached from the outside and thus resulting in just wandering into a
seemingly random iron giant.

Also also, put back a Depths scaling of one vault, but replace the
previous now-taboo iron giant with a lordly juggernaut instead.