Fix a Legendary Destruction freeze.
When the target was within the range of Legendary Destruction, but not
within the range of any of the possible spells, setup_mons_cast would
spin in a loop trying and failing to pick a spell with larger range.

Now that all the component spells have the same range, remove that loop
and adjust LD's range to be the same.