Rescale Lair monster list (elliptic)
To make the new Lair:1 closer to old Lair:1 (instead of being based on old
Lair:3 as with the initial implementation of the level cuts), so as not to
make the jump from D to Lair too large. Total XP in Lair remains rougly the
same as with the previous scaling. The old Lair:1-8 spawns are now roughly
compressed into the new Lair:1-6 (with a bunch of rounding and some arbitrary
fiddling, further adjustments from those more familiar with mon-pick are very
welcome!).