Flatten sack of spiders and box of beasts self-destruct chances (minmay)
The new odds are 1 in 4 regardless of Evo skill. The reasoning is that
we don't want the strength of these items to scale so heavily with Evo
skill -- the strength of the spiders/beasts released already scales
heavily, and we don't necessarily want the high end to be as strong as
it was.