Display Hepliaklqana ancestors at the top of the monster list
Unlike other allies which are deliberately listed below all hostiles, since
this is a case where the player is likely to have only one ally whose health
and status they want to keep track of, especially in larger fights. Always
showing it in the monster list fixes this requiring lots of xv-ing in these
situations in console (at the cost of one less monster list spot for hostiles).