With Control Winds gone, replace cloud mage Cloud Cone / Blink with Tornado
Higher tier clouds on monsters is a little neat, and the cloud mage would make
lots of them but be fine even when blinking around with the cloud redirecting
status. With the immunity gone due to its superfluous main use on spriggan air
mages, it's rather easy for the cloud mage to die to their own clouds now.

Tornado is certainly a bit much of a replacement, but it's simpler than the
overall cloud engine and the only other use on monsters as stands is Lom Lobon.
The addition of emergency flight avoids potential drowning deaths from it,
anyway.