Players' non-opaque clouds instavanish outside LOS
They previously dissapated at 4x speed to discourage abuses, which
was (a) ineffective and (b) hard to notice unless you already knew
about it. ("Spoilery!") Better to make a hard-and-fast rule.

Opaque clouds are exempt so that e.g. large amounts of fog or
steam doesn't end up insta-dissipating; possibly this should be
revisited later.