Convert an old infiniplex wizlab_nikola submission (#6667) into a Zot vault
Nobody ever did pick up the problems it had with trying to drop its random
hostile ball lightning summons, even as over the years more electric enemies
provide more of a viable thematic monster set. I was also never fond of the
way it broke up the other wizlabs corresponding to titles, spells, and unands
by corresponding to a unique instead. Infiniplex eventually disappeared for
quite the while now, anyway.

So I've converted the layout into a large and fancy Zot vault, without the lua
gimmick and with a bit more hypothetical threat. There's still a few awkward
elements to it- due to using grates one can blink into the tornado turbines
(which have some hatch safe spaces just in case) and currently the
flavour-spawn hostile ball lightnings are bugged to try and explode even on the
other side of grates- but it's still otherwise a nice and fancy vault in the
company of many, many, many other Zot gimmick vaults.