Somewhat flatten storm cloud hits / damage
Storm clouds make for odd play in rare monster hands or to risk entering with
the rod of clouds: 1/4th of the time they hit for four times the damage of fire
and frost clouds (24-87 elec damage for monsters, 40 to 131 elec damage for
players!), which is interesting as a non-debilitating cloud appropiate for
lategame and extended but also with an unfortunately randomized and spoilery
degree of large spikes of damage.

To make them less absurd and more regulated yet still recognizable, storm
clouds now strike monsters for 21-61 and players for 35-80 damage 1/3 of the
time and monsters (3.5x bottom of standard damage, 2.5x top of standard
damage). Nikola's chain lightning or the average late-game conjuration can do
worse, and it's still a gamble to sit around in the clouds. Cloud strikes to
non-occupied tiles are still a 1/4 chance, due to the very nature of that
being plenty of message spam as is (though this can be altered).