Remove Hellwings' Cigotuvi's Embrace
Didn't do much, would be probably be an issue if it did (as with
other monster spells that use corpses). I had a plan for them to
get homonuculus buddies that would synergize with them (hit with
AF_VULN to let their Teleport Other trigger, die to power their
master's Cigotuvi's Embrace), but the former spell was removed and
the latter doesn't really interact super well with AF_SWOOP.

Ironically, this means they're back to being basically blue devils.