Player can no longer tear web trap on escape
Web traps having a chance to permanently tear on player escape incentives
dancing in and out of trap until it is destroyed, if the trap is in a
strategically useful spot (e.g., a chokepoint). Monster interaction with webs is
untouched by this change, so the player can still manipulate monster pathing to
destroy web traps. Discussed at length at
https://crawl.develz.org/tavern/viewtopic.php?f=8&t=19822