Make Bind Souls non-permanent (PleasingFungus).
The enchantment wears off after a 10-30 turns. The messaging was changed
to not trigger per-monster to be less spammy, and the casting rate was
increased to compensate.

The original motivation for making it permanent was to discourage kiting
the monster around until it wore off once you killed the necromancer;
however, you could kite the spectre around anyway if you didn't want to
fight it, since it's summoned...