Simplify (and improve accuracy of a chance).
The chance of considering a teleport from teleportitis after a player
turn was intended to be 1/(100/AUT) = AUT/100, but random rounding gave
a different result.

Consider the case of a 9-aut action.  We considered a teleport if
one_chance_in(div_rand_round(100, 9)) was true.  This argument to
one_chance_in has the distribution:

   div_rand_round(100, 9) = { 8/9 chance of 11, 1/9 chance of 12 }

so the probability of considering a teleport in this situation was
actually (8/9 * 1/11) + (1/9 * 1/12) = 0.090067, slightly higher than
the desired 9/100 = 0.09.

Instead, use the simpler x_chance_in_y(9, 100).  Durations that evenly
divide 100 AUT are unaffected.

Also, write 100 as 10 * BASELINE_DELAY, to clarify the base 1/10 chance.