Make poison application messages announce maximum poison
Now if you would add a level of poison to a creature at maximum poison,
instead of no message, there will be a message about the monster being 'as
poisoned as possible'. The message will also print if the monster reached
maximum poisoning level in this application of poison.

This also introduces a constant for maximum enchantment degree. I'm sure
there are lots of places where this could be used, but I don't feel like
hunting them down. If you see something, say something!