Thorough Temple edits/clean-up, some new maps, altar average 12.15 -> 14.41
Temple.des is a mess. Outdated, rough section organization and notes based
on themes and colours, early altar balance shifting against every new god,
vast differences in sizes. The vast majority of weight goes towards 12
altars, despite how the versions have collected 22 guaranteed gods in D.

With a lot of temple map editing and variation compression, plus some
header clean-up, this bumps up the minimum of altars to 9 (from 6 with one 5)
and the max to 21 (mostly 18 with a few 20s), and moves from an average of
12.154 altars in each temple to 14.409. Some maps have gained some more
randomization and decoration to them in the process, while some other larger
maps were trimmed or made more interconnected to address auto-explore length
concerns.

Two of evktalo's old and removed temple designs were refurbished and brought
out, and I threw together six of my own designs in here as well. Their
presence helps with the raw map count compression versus a flattening of
vault weights (10 for a fixed count and +4 for variable counts per each
range met, halved weights for either if they meet certain criteria).
Beyond a much lowered appearance rate for the original circular and lava
triangle maps and a bit more prominence for a handful of crystal layouts,
most of the maps haven't changed in appearance rate by more than a few 0.1%s.

XXX: The specific altar ranges here were planned for a while to dynamically
adjust vaults chances to future-proof altar averages versus newly-added
vaults (with the presumption the guaranteed god count is going to stay at
a rough 21 to 24 for the next long while). Something like a 25% chance for
9 to 12 altars, a 35% chance for 13 to 15, a 30% chance for 16 to 18, and
a 10% chance for 21. This could be done on the C side choosing given altar
counts against those weighted chances, though about ten more maps for the
middle ranges ontop of the current forty-nine would be ideal to reach an
even ground for temples chances of all given ranges.